feat: extract shared utilities from macOS ChatView.swift
