+++ clients/shared/Utilities/FileDisplayHelpers.swift
@@ -1,0 +1,24 @@
+import Foundation
+
+/// Shared file display helpers for icon selection and size formatting.
+public enum FileDisplayHelpers {
+
+    /// Returns an SF Symbol name appropriate for the given MIME type.
+    public static func fileIcon(for mimeType: String) -> String {
+        if mimeType.hasPrefix("text/") { return "doc.text.fill" }
+        if mimeType == "application/pdf" { return "doc.fill" }
+        if mimeType.contains("zip") || mimeType.contains("archive") { return "doc.zipper" }
+        if mimeType.contains("json") || mimeType.contains("xml") { return "doc.text.fill" }
+        return "doc.fill"
+    }
+
+    /// Formats a base64-encoded payload length into a human-readable file size.
+    public static func formattedFileSize(base64Length: Int) -> String {
+        let bytes = base64Length * 3 / 4
+        if bytes < 1024 { return "\(bytes) B" }
+        let kb = Double(bytes) / 1024
+        if kb < 1024 { return String(format: "%.1f KB", kb) }
+        let mb = kb / 1024
+        return String(format: "%.1f MB", mb)
+    }
+}
